+++ static/app/utils/replays/hooks/useExtractDomNodes.tsx
